moved some test files to models package
