+++ test/uk/gov/hmrc/offpayroll/models/OffpayrollWebflowSpec.scala
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.offpayroll
+package uk.gov.hmrc.offpayroll.models
 
 import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{FlatSpec, Matchers}
 import uk.gov.hmrc.offpayroll.models.DecisionBuilder.Interview
-import uk.gov.hmrc.offpayroll.models.{Cluster, Element, OffPayrollWebflow, Webflow}
 
 /**
   * Created by peter on 05/12/2016.
